+++ lib/slack-ruby-bot-server.rb
@@ -2,6 +2,7 @@ require 'async/websocket'
 
 require 'grape-swagger'
 require 'slack-ruby-bot'
+require 'slack-ruby-bot-server/service'
 require 'slack-ruby-bot-server/server'
 require 'slack-ruby-bot-server/config'
 
@@ -13,4 +14,3 @@ require "slack-ruby-bot-server/config/database_adapters/#{SlackRubyBotServer::Co
 
 require 'slack-ruby-bot-server/api'
 require 'slack-ruby-bot-server/app'
-require 'slack-ruby-bot-server/service'