+++ lib/slack-ruby-bot-server/config.rb
@@ -3,10 +3,12 @@ module SlackRubyBotServer
     extend self
 
     attr_accessor :server_class
+    attr_accessor :service_class
     attr_accessor :database_adapter
 
     def reset!
       self.server_class = SlackRubyBotServer::Server
+      self.service_class = SlackRubyBotServer::Service
       self.database_adapter = if defined?(::Mongoid)
                                 :mongoid
                               elsif defined?(::ActiveRecord)